Added support for using the simulator
