+++ example/tranversing_address_map/dumping_register_state_to_json_file.py
@@ -2,47 +2,11 @@ import json
 from typing import Union
 
 from chip_with_registers.reg_model.chip_with_registers import chip_with_registers_cls
+from chip_with_registers.sim.chip_with_registers import chip_with_registers_simulator_cls
 
 from chip_with_registers.lib import NormalCallbackSet, RegReadOnly, RegReadWrite, \
     MemoryReadOnly, MemoryReadWrite, RegFile, AddressMap, RegReadOnlyArray, RegReadWriteArray, \
     AddressMapArray, RegFileArray, MemoryReadOnlyArray, MemoryReadWriteArray
-
-
-# dummy functions to demonstrate the class
-def read_addr_space(addr: int, width: int, accesswidth: int) -> int:
-    """
-    Callback to simulate the operation of the package, everytime the read is called, it return
-    an integer value of 0
-
-    Args:
-        addr: Address to write to
-        width: Width of the register in bits
-        accesswidth: Minimum access width of the register in bits
-
-    Returns:
-        value inputted by the used
-    """
-    assert isinstance(addr, int)
-    assert isinstance(width, int)
-    assert isinstance(accesswidth, int)
-    return int(0)
-
-
-def write_addr_space(addr: int, width: int, accesswidth: int, data: int) -> None:
-    """
-    Callback to simulate the operation of the package, everytime the read is called, it will
-    request the user input the value to be read back.
-
-    Args:
-        addr: Address to write to
-        width: Width of the register in bits
-        accesswidth: Minimum access width of the register in bits
-        data: value to be written to the register
-
-    Returns:
-        None
-    """
-    raise NotImplementedError('No register write occur in this example')
 
 
 class RegisterDumper:
@@ -172,8 +136,9 @@ if __name__ == '__main__':
 
     # create an instance of the address map with the simulated callback necessary to demonstrate
     # the example
-    dut = chip_with_registers_cls(callbacks=NormalCallbackSet(read_callback=read_addr_space,
-                                                              write_callback=write_addr_space))
+    sim = chip_with_registers_simulator_cls(0)
+    dut = chip_with_registers_cls(callbacks=NormalCallbackSet(read_callback=sim.read,
+                                                              write_callback=sim.write))
 
     # generate an instance of the RegisterDumper and write the registers to a file
     reg_dumper = RegisterDumper(dut)
